+++ graalpython/com.oracle.graal.python.cext/include/hpy/universal/autogen_trampolines.h
@@ -645,7 +645,7 @@ HPyAPI_FUNC HPy HPyCapsule_New(HPyContext *ctx, void *pointer, const char *name,
 }
 
 HPyAPI_FUNC void *HPyCapsule_Get(HPyContext *ctx, HPy capsule, _HPyCapsule_key key, const char *name) {
-     return ctx->ctx_Capsule_Get ( ctx, UNWRAP(capsule), key, name );
+     return ctx->ctx_Capsule_Get ( ctx, UNWRAP(capsule), (int32_t)key, name );
 }
 
 HPyAPI_FUNC int HPyCapsule_IsValid(HPyContext *ctx, HPy capsule, const char *name) {
@@ -653,7 +653,7 @@ HPyAPI_FUNC int HPyCapsule_IsValid(HPyContext *ctx, HPy capsule, const char *nam
 }
 
 HPyAPI_FUNC int HPyCapsule_Set(HPyContext *ctx, HPy capsule, _HPyCapsule_key key, void *value) {
-     return ctx->ctx_Capsule_Set ( ctx, UNWRAP(capsule), key, value );
+     return ctx->ctx_Capsule_Set ( ctx, UNWRAP(capsule), (int32_t)key, value );
 }
 
 HPyAPI_FUNC HPy HPy_FromPyObject(HPyContext *ctx, cpy_PyObject *obj) {
